support for es 1.2
